+++ src/core/outcome_spaces.jl
@@ -12,11 +12,12 @@ The supertype for all outcome space implementation.
 
 In ComplexityMeasures.jl, an outcome space defines a set of possible outcomes
 ``\\Omega = \\{\\omega_1, \\omega_2, \\ldots, \\omega_L \\}`` (some form of
-discretization). In the literature, the outcome space is often called an "alphabet",
+discretization). In the literature, the outcome space is often also called an "alphabet",
 while each outcome is called a "symbol" or an "event".
 
 An outcome space also defines a set of rules for mapping input data to
-to each outcome ``\\omega_i`` (i.e. [encoding](@ref encodings)/discretizing).
+to each outcome ``\\omega_i``, a processes called _encoding_ or _symbolizing_ or
+_discretizing_ in the literature (see [encodings](@ref encodings)).
 Some [`OutcomeSpace`](@ref)s first apply a transformation, e.g. a delay embedding, to
 the data before discretizing/encoding, while other [`OutcomeSpace`](@ref)s
 discretize/encode the data directly.